LayoutTests/fast/dom/Document/adoptNode-quirks-mismatch2-expected.txt

Tests that Document.adoptNode() properly deals with the documents having a different quirks mode

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

MyClass

MyId

PASS window.getComputedStyle(container.lastChild).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(container.lastChild).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
